requirements.txt
# Web Framework
fastapi==0.117.1
pydantic==2.5.3
uvicorn==0.38.0
websockets==13.1

# Data Generation
pyyaml==6.0.1
faker==38.0.0
exrex==0.12.0

# Testing
pytest==9.0.1
pytest-asyncio==1.3.0
pytest-cov==6.0.0
requests==2.31.0
httpx
websocket-client

# Observability
structlog==25.5.0
prometheus-client==0.23.1

# Persistence
redis==7.1.0
asyncpg==0.31.0

# Utilities
nanoid==2.0.0
uvloop==0.22.1
orjson==3.11.4

# Development
ruff==0.14.9
